* feat(Docs): Add a framework for including docs in-app.
* Clean up the main App.tsx file a bit.

* feat(Docs): Add a framework for including docs in-app.

Docs are stored in the client/src/docs folder and are organized
into sections using the folder structure. Doc pages support frontmatter
for the `title` and `order` which they appear in the navigation sidebar.

Docs support MDX, which means they can import JSX components from
elsewhere in the app, which could be useful for interactive demos.

It requires a patch to one of the plugin packages to allow frontmatter to
be read from the routes file.