Component refactor (#48)
* Update some of the docs.

* fix(ECS): Improve the API for defining components

As I was writing docs, I realized that the component implementation
is complicated and convoluted. This removes the silly static
defaults property in favor of adding defaults directly to the
properties. This does require making it so the "constructor" stuff
is handled by a separate "init" function, since default properties
override properties defined in the parent class constructor. Ah well,
this is still better, and the .init() API is unlikely to be used outside
of the entity class anyway.

* Fixes a few things.

* Fix a formatting issue.