Lots of changes to plugins.
Begin storing data as YAML, to make it easier to edit by hand.

Refactor the db-fs package to use class inheritance rather than wrapping objects.
This provides a neater experience, with fewer hacks for the end user,
although the implementation itself is full of hacks ;)

Created a function that can check existing names to make sure names
aren't being duplicated. This makes it possible to use names as IDs
while still keeping them unique.

Add a Snapshot button to the development version of the app. This
serves as the save button, which makes it so you can restart the server
to get it back to an expected state, but click the button to save the
state again.

Started writing plugins to the file system. Yay!