Plugins 2 (#76)
* Add the ability for Thorium Nova to host plugin assets. Closes #64

* Starting on the plugin editing UI

* Initial UI for the plugin configuration page.

* Basic plugin definition. Closes #58

* Implement the NetRequest system

This makes it possible to have clients request any data
using parameters. It's basically a glorified real-time live data
REST API over websockets, really.

The cool thing about it is that it's implemented with React Suspense
so to the developer, it behaves as if you were just calling a function.

Closes #59

* UPdate the properties of plugins. Closes #60

* Starting on asset upload.
Replaced the WebSocket netsend API with HTTP,
since it makes it easier to include the uploaded files
inline.

Todo: Create a utility function for moving files
wherever you want, most likely placing them inside the
assets folder. The good news is that the input function
is able to decide whatever they want to do with the file,
which is nice.

* Fix an issue

* Add support for uploading a cover image to a plugin.